Extract normalizeTemplateIdentifier to a trait
